requirements.txt
langgraph==1.0.8
langchain-google-genai==4.2.0
langchain-core==1.2.11
python-dotenv==1.2.1
pytest==9.0.2

